Add debian directory
Add debian directory so that Debian packages can be built from
the source.  The debian directory is the same as the one used by
Ubuntu Cloud Archive for Havana, intended to build packages to run
on Precise.  In this commit, Ubuntu Cloud Archive has not updated
its code to reflect the 'quantum'->'neutron' name change, so it
does not build.  This commit is meant to simply to initialize.

Signed-off-by: Ryu Ishimoto <ryu@midokura.com>